Renamed package sjchat.backend.** to just sjchat.**.
