+++ src/main/java/sjchat/messages/MessageServiceServer.java
@@ -1,4 +1,4 @@
-package sjchat.backend.messages;
+package sjchat.messages;
 
 import java.io.IOException;
 import java.util.Random;
@@ -6,7 +6,7 @@ import java.util.Random;
 import io.grpc.Server;
 import io.grpc.ServerBuilder;
 import io.grpc.stub.StreamObserver;
-import sjchat.backend.users.User;
+import sjchat.users.User;
 
 public class MessageServiceServer {
   private Server server;
