+++ src/main/java/sjchat/restapi/Application.java
@@ -1,4 +1,4 @@
-package sjchat.backend.restapi;
+package sjchat.restapi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
