+++ src/main/proto/message-service.proto
@@ -5,7 +5,7 @@ package sjchat.messages;
 import "user-service.proto";
 
 option java_multiple_files = true;
-option java_package = "sjchat.backend.messages";
+option java_package = "sjchat.messages";
 option java_outer_classname = "MessageServiceProto";
 
 service MessageService {
