+++ src/main/proto/user-service.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 package sjchat.users;
 
 option java_multiple_files = true;
-option java_package = "sjchat.backend.users";
+option java_package = "sjchat.users";
 option java_outer_classname = "UserServiceProto";
 
 service UserService {
